Fix & Improvement: Apply to ALL the jobs
Fixed the issue where the script would click on a link that would redirect it towards the job-post page. Now clicks the image of the post to avoid being redirected.

Implemented the ability to go to the next page. Code will go through all of the available pages and apply to all the jobs.